+++ src/common/components/Notification/styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const NotificationWrapper = styled.div``;
+
+export const NotificationText = styled.p``;
