+++ src/common/components/Toolbar/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import useAuth from 'src/common/hooks/useAuth';
 import AuthModal from 'src/common/components/AuthModal';
-import styled from 'styled-components';
+import { Header, Title } from './styles';
 
 interface ToolbarProps {}
 
@@ -18,34 +18,18 @@ export default function Toolbar({ children }: React.PropsWithChildren<ToolbarPro
 	}
 
 	return (
-		<>
-			<Header>
-				<Title>Public forum</Title>
-				{authUser ? (
-					<>
-						<div>{authUser.alias}</div>
-						<button onClick={signOutUser}>Logout</button>
-					</>
-				) : (
-					<button onClick={showAuthModal}>Sign In</button>
-				)}
-				{children}
-			</Header>
+		<Header>
+			<Title>Public forum</Title>
+			{authUser ? (
+				<>
+					<div>{authUser.alias}</div>
+					<button onClick={signOutUser}>Logout</button>
+				</>
+			) : (
+				<button onClick={showAuthModal}>Sign In</button>
+			)}
+			{children}
 			<AuthModal active={isAuthVisible} onClose={closeModal} />
-		</>
+		</Header>
 	);
 }
-
-const Header = styled.header`
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	background-color: #45086e;
-	color: white;
-	padding: 1rem 2rem;
-`;
-
-const Title = styled.h1`
-	margin: 0;
-	font-size: 1.2rem;
-`;
