+++ src/common/components/Toolbar/styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Header = styled.header`
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	background-color: #45086e;
+	color: white;
+	padding: 1rem 2rem;
+`;
+
+export const Title = styled.h1`
+	margin: 0;
+	font-size: 1.2rem;
+`;
